refactor: Update VideoControllerCubit to use VideoModel instead of VideoPlayerController
This commit refactors the `VideoControllerCubit` to use the `VideoModel` class instead of the `VideoPlayerController` class. The `setController` method has been updated to no longer accept a `VideoPlayerController` parameter. Instead, it now sets the state to `VideoControllerInitial` without a controller. The `play` method now accepts a `VideoModel` parameter and emits a `VideoControllerPlaying` state with the provided model. The `pause` method no longer disposes the controller and emits a `VideoControllerPaused` state. This change improves the flexibility and reusability of the `VideoControllerCubit` by decoupling it from the specific implementation of the video player.